Finalizando responsividade, adicionando logica para quando for celular passando o isMobile
