+++ src/components/Banner/banner.scss
@@ -1,0 +1,7 @@
+.banner{
+   width: 100%;
+
+   img{
+    width: 100%;
+   }
+}
