+++ src/components/Banner/index.tsx
@@ -1,14 +1,21 @@
 import { Swiper, SwiperSlide } from 'swiper/react';
-
 import BannerImg01 from '../../assets/banners/banner01.png';
 import BannerImg02 from '../../assets/banners/banner02.png';
 import BannerImg03 from '../../assets/banners/banner03.png';
+import BannerImg01Mobile from '../../assets/banners/banner01Mobile.png';
+import BannerImg02Mobile from '../../assets/banners/banner02Mobile.png';
+import BannerImg03Mobile from '../../assets/banners/banner03Mobile.png';
 import { Pagination, Autoplay } from 'swiper';
+import './banner.scss';
 
-export const Banner = () => {
+interface BannerProps {
+  isMobile: boolean;
+}
+
+export const Banner = ({ isMobile }: BannerProps) => {
   return (
     <>
-      <section>
+      <section className="banner">
         <Swiper
           slidesPerView={1}
           modules={[Pagination, Autoplay]}
@@ -20,13 +27,22 @@ export const Banner = () => {
           }}
         >
           <SwiperSlide>
-            <img src={BannerImg01} alt="" />
+            <img
+              src={isMobile ? BannerImg01Mobile : BannerImg01}
+              alt="Banner 1"
+            />
           </SwiperSlide>
           <SwiperSlide>
-            <img src={BannerImg02} alt="" />
+            <img
+              src={isMobile ? BannerImg02Mobile : BannerImg02}
+              alt="Banner 2"
+            />
           </SwiperSlide>
           <SwiperSlide>
-            <img src={BannerImg03} alt="" />
+            <img
+              src={isMobile ? BannerImg03Mobile : BannerImg03}
+              alt="Banner 3"
+            />
           </SwiperSlide>
         </Swiper>
       </section>
