+++ src/components/FeaturedProduct/index.tsx
@@ -1,13 +1,17 @@
 import './featured.scss';
 import { Product } from './Product';
 
-export const FeaturedProduct = () => {
+interface FeaturedProductProps {
+  isMobile: boolean;
+}
+
+export const FeaturedProduct = ({ isMobile }: FeaturedProductProps) => {
   return (
     <>
       <section className="featuredProduct">
         <div className="featuredProduct__div">
           <h2>Mais Vendidos</h2>
-          <Product />
+          <Product isMobile={isMobile} />
         </div>
       </section>
     </>
