+++ src/components/Footer/footer.scss
@@ -1,7 +1,13 @@
 .footer{
     background-color: #000;
     width: 100%;
-    height: 270px;
+    height: 320px;
+    padding: 3rem;
+
+    @media screen and (max-width: 48rem) {
+        padding: 1rem;
+        height: 470px;
+    }
 
     &__container{
         max-width: 1280px;
@@ -14,6 +20,7 @@
         align-items: center;
         justify-content: center;
         gap: 2rem;
+      
     }
 
     button {
@@ -26,10 +33,10 @@
             font-weight: 700;
             width: 250px;
             transition: background-color 0.3s ease;
-            margin-top: 20px;
             border-radius: 0.3125rem;
             height: 3rem;
             transition: opacity 0.12s ease;
+            line-height: 1.2rem;
     
             &:hover {
                 background-color: #cfcfcf;
@@ -82,5 +89,14 @@
                 width: 156px;
              }
     
+         }
+
+         &__mobile-footer{
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            gap: 1rem;
+            margin-top: 3rem;
+
          }
 }
