+++ src/components/Header/header.scss
@@ -8,13 +8,9 @@
     padding: 20px;
     font-weight: 400;
 
-    // @media screen and (max-width: 48rem) {
-    //     gap: 4px;
-    //     max-width: 48px;
-    //     margin: auto;
-    //     padding: 20px;
-    //     font-weight: 400;
-    //   }
+    @media screen and (max-width: 48rem) {
+      width: 100%;
+     }
 
     div{
       display: flex;
@@ -27,7 +23,9 @@
     }
 
     span{
-          display: inline;
+          display: flex;
+          align-items: center;
+          gap: 7px;
         @media screen and (max-width: 48rem) {
                 display: none;
         }
@@ -74,20 +72,22 @@
   &__inputSearch {
     display: flex;
     align-items: center;
-    width: 44rem;
+    width: 100%;
     border-bottom: 1px solid #7a7a7a;
     padding-bottom: 0.25rem;
 
-      @media screen and (max-width: 48rem) {
-        width: 22rem;
-        margin-top: 1rem;
-      }
 
     input {
-      width: 100%;
+      width: 44rem;
       color: #7a7a7a;
       font-size: 0.8125rem;
       font-weight: 400;
+
+      @media screen and (max-width: 48rem) {
+        width: 350px;
+        margin-top: 1rem;
+      }
+
     }
 
     img {
