+++ src/styles/App.css
@@ -43,9 +43,6 @@
 }
 
 
-
-
-
 .swiper-button-next {
   right: -20px;
 }
